# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Foundry - Decentralized AI Model Fine-Tuning Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FineTuneJobStatusProperty =
  (typeof FineTuneJobStatusProperty)[keyof typeof FineTuneJobStatusProperty];

export const FineTuneJobStatusProperty = {
  pending: "pending",
  uploading: "uploading",
  training: "training",
  completed: "completed",
  failed: "failed",
} as const;

export interface FineTuneJob {
  id: number;
  creatorWallet: string;
  /** @nullable */
  jobIdOn0g?: string | null;
  /** @nullable */
  datasetRootHash?: string | null;
  /** @nullable */
  datasetOgTxHash?: string | null;
  /** @nullable */
  datasetOgExplorerUrl?: string | null;
  baseModel: string;
  modelName: string;
  description: string;
  category: string;
  status: FineTuneJobStatusProperty;
  /** @nullable */
  progressPct?: number | null;
  /** @nullable */
  modelRootHash?: string | null;
  /** @nullable */
  modelOgExplorerUrl?: string | null;
  /** @nullable */
  nftTokenId?: string | null;
  /** @nullable */
  nftOgChainTxHash?: string | null;
  /** @nullable */
  nftOgExplorerUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface FineTuneJobStatus {
  id: number;
  status: string;
  /** @nullable */
  progressPct?: number | null;
  /** @nullable */
  jobIdOn0g?: string | null;
  /** @nullable */
  modelRootHash?: string | null;
  /** @nullable */
  modelOgExplorerUrl?: string | null;
  /** @nullable */
  nftTokenId?: string | null;
}

export type CreateFineTuneJobBodyBaseModel =
  (typeof CreateFineTuneJobBodyBaseModel)[keyof typeof CreateFineTuneJobBodyBaseModel];

export const CreateFineTuneJobBodyBaseModel = {
  "Qwen25-05B-Instruct": "Qwen2.5-0.5B-Instruct",
  "Qwen3-32B": "Qwen3-32B",
} as const;

export type CreateFineTuneJobBodyCategory =
  (typeof CreateFineTuneJobBodyCategory)[keyof typeof CreateFineTuneJobBodyCategory];

export const CreateFineTuneJobBodyCategory = {
  "customer-support": "customer-support",
  "creative-writing": "creative-writing",
  "code-assistant": "code-assistant",
  finance: "finance",
  medical: "medical",
  legal: "legal",
  other: "other",
} as const;

export interface CreateFineTuneJobBody {
  creatorWallet: string;
  baseModel: CreateFineTuneJobBodyBaseModel;
  modelName: string;
  description: string;
  category: CreateFineTuneJobBodyCategory;
  datasetDescription: string;
  samplePrompt: string;
  sampleOutput: string;
  licensePriceUsd: number;
  /** JSONL content of the training dataset */
  datasetContent: string;
}

export interface Model {
  id: number;
  /** @nullable */
  jobId?: number | null;
  /** @nullable */
  nftTokenId?: string | null;
  /** @nullable */
  ogChainTxHash?: string | null;
  /** @nullable */
  ogExplorerUrl?: string | null;
  creatorWallet: string;
  name: string;
  description: string;
  category: string;
  baseModel: string;
  datasetDescription: string;
  samplePrompt: string;
  sampleOutput: string;
  licensePriceUsd: number;
  isListed: boolean;
  inferenceCount: number;
  licenseCount: number;
  /** @nullable */
  modelRootHash?: string | null;
  /** @nullable */
  datasetRootHash?: string | null;
  createdAt: string;
}

export interface UpdateModelBody {
  name?: string;
  description?: string;
  licensePriceUsd?: number;
  isListed?: boolean;
  samplePrompt?: string;
  sampleOutput?: string;
}

export interface ListModelBody {
  licensePriceUsd: number;
  creatorWallet: string;
}

export interface License {
  id: number;
  modelId: number;
  buyerWallet: string;
  /** @nullable */
  ogPaymentTxHash?: string | null;
  /** @nullable */
  ogExplorerUrl?: string | null;
  activeUntil: string;
  createdAt: string;
  model: Model;
}

export interface PurchaseLicenseBody {
  modelId: number;
  buyerWallet: string;
  durationDays: number;
}

export interface InferModelBody {
  prompt: string;
  callerWallet: string;
  systemPrompt?: string;
}

export interface InferModelResponse {
  modelId: number;
  response: string;
  /** @nullable */
  teeProvider?: string | null;
  /** @nullable */
  teeAttestationRef?: string | null;
  processingMs: number;
  calledAt: string;
}

export type ActivityEventEventType =
  (typeof ActivityEventEventType)[keyof typeof ActivityEventEventType];

export const ActivityEventEventType = {
  model_trained: "model_trained",
  model_listed: "model_listed",
  license_purchased: "license_purchased",
  inference_run: "inference_run",
  job_started: "job_started",
} as const;

export interface ActivityEvent {
  id: number;
  eventType: ActivityEventEventType;
  /** @nullable */
  modelId?: number | null;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  actorWallet?: string | null;
  /** @nullable */
  ogExplorerUrl?: string | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export interface PlatformStats {
  totalModels: number;
  totalFineTuneJobs: number;
  totalInferenceCalls: number;
  totalLicenses: number;
  totalDatasetStorageGb: number;
  featuredModels: Model[];
}

export interface CreatorStats {
  totalJobs: number;
  completedJobs: number;
  listedModels: number;
  totalLicenses: number;
  totalInferenceCalls: number;
  estimatedEarningsUsd: number;
  jobs: FineTuneJob[];
  models: Model[];
}

export type ListFineTuneJobsParams = {
  creatorWallet?: string;
};

export type ListModelsParams = {
  baseModel?: string;
  category?: string;
  sort?: ListModelsSort;
  creatorWallet?: string;
};

export type ListModelsSort =
  (typeof ListModelsSort)[keyof typeof ListModelsSort];

export const ListModelsSort = {
  popular: "popular",
  newest: "newest",
  cheapest: "cheapest",
} as const;

export type ListLicensesParams = {
  buyerWallet?: string;
};

export type GetActivityParams = {
  limit?: number;
};

export type GetCreatorStatsParams = {
  creatorWallet: string;
};