# lib/api-zod/src/generated/types/createFineTuneJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Foundry - Decentralized AI Model Fine-Tuning Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFineTuneJobBodyBaseModel } from "./createFineTuneJobBodyBaseModel";
import type { CreateFineTuneJobBodyCategory } from "./createFineTuneJobBodyCategory";

export interface CreateFineTuneJobBody {
  creatorWallet: string;
  baseModel: CreateFineTuneJobBodyBaseModel;
  modelName: string;
  description: string;
  category: CreateFineTuneJobBodyCategory;
  datasetDescription: string;
  samplePrompt: string;
  sampleOutput: string;
  licensePriceUsd: number;
  /** JSONL content of the training dataset */
  datasetContent: string;
}