# lib/api-zod/src/generated/types/fineTuneJobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Foundry - Decentralized AI Model Fine-Tuning Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface FineTuneJobStatus {
  id: number;
  status: string;
  /** @nullable */
  progressPct?: number | null;
  /** @nullable */
  jobIdOn0g?: string | null;
  /** @nullable */
  modelRootHash?: string | null;
  /** @nullable */
  modelOgExplorerUrl?: string | null;
  /** @nullable */
  nftTokenId?: string | null;
}