# lib/api-zod/src/generated/types/inferModelResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Foundry - Decentralized AI Model Fine-Tuning Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface InferModelResponse {
  modelId: number;
  response: string;
  /** @nullable */
  teeProvider?: string | null;
  /** @nullable */
  teeAttestationRef?: string | null;
  processingMs: number;
  calledAt: Date;
}