# lib/db/src/schema/licenses.ts
import { pgTable, text, serial, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const licensesTable = pgTable("licenses", {
  id: serial("id").primaryKey(),
  modelId: integer("model_id").notNull(),
  buyerWallet: text("buyer_wallet").notNull(),
  ogPaymentTxHash: text("og_payment_tx_hash"),
  ogExplorerUrl: text("og_explorer_url"),
  activeUntil: timestamp("active_until", { withTimezone: true }).notNull(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertLicenseSchema = createInsertSchema(licensesTable).omit({
  id: true,
  createdAt: true,
});
export type InsertLicense = z.infer<typeof insertLicenseSchema>;
export type License = typeof licensesTable.$inferSelect;